No longer support USR1 signal in parent process
Although the use of USR1 signal in parent process was convenient for
automatically reload the Pearl configuration, it turned out that there
are significant drawbacks with it. In case the USR1 signal is not caught by the
parent (i.e. non interactive shells), the parent process will be terminated
immediately by the Pearl process.
In case a new trap in USR1 overrides the one defined by Pearl,
the parent process would behave unexpectedly.

In this code change we introduce a simple function `pearl-source`
available on any shell such that user can call it any time to reload
the Pearl configuration. This is not always required but depends
whether the variables/functions involve the current shell where the
package install/update occurred (i.e. a new variable defined in `config.sh`
and the current shell is a bash or zsh). Alternatively, user can always
create a new shell and the package resources will be available as
expected.